+++ uefi-raw/CHANGELOG.md
@@ -1,6 +1,9 @@
 # uefi-raw - [Unreleased]
 
 ## Added
+- Added `ResetNotification`.
+
+## Added
 - Added `TimestampProtocol`.
 
 # uefi-raw - 0.5.1 (2024-03-17)